% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gen_p}
\alias{gen_p}
\title{generate conditional probability of U}
\usage{
gen_p(
  data,
  inter = FALSE,
  p.U = 0.5,
  outcome,
  time,
  m,
  t,
  covariates,
  b.y,
  b.m,
  b.t,
  scale.m,
  scale.t,
  scale.u,
  sigma.u = 1,
  iter_num = 20,
  lower_u = -10,
  upper_u = 10
)
}
\arguments{
\item{data}{data frame including time, outcome, m, t, and covariates}

\item{inter}{a logical value indicating the existence of exposure-mediator interaction}

\item{p.U}{marginal probability of unmeasured confounder}

\item{outcome}{survival outcome}

\item{time}{survival time}

\item{m}{mediator}

\item{t}{exposure}

\item{covariates}{the vector of covariates}

\item{b.y, b.m, b.t}{the coefficients between U and exposure, mediator, and survival outcome}

\item{scale.m, scale.t, scale.u}{scale of mediator, exposure and U, binary or continuous}

\item{sigma.u}{The standard deviation of the prior distribution of U if it is continuous. The default is 1.}

\item{iter_num}{number of iteration}

\item{lower_u, upper_u}{lower and upper bound of the integration interval for U if it is continuous.}
}
\value{
U and its conditional probability (if U is binary)
}
\description{
generate conditional probability of U
}
