% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gen_p}
\alias{gen_p}
\title{generate conditional probability of U}
\usage{
gen_p(
  data,
  inter = FALSE,
  p.U,
  outcome,
  time,
  m,
  t,
  covariates,
  b.y,
  b.m,
  b.t,
  scale.m,
  scale.t,
  iter_num = 20
)
}
\arguments{
\item{data}{data frame including time, outcome, m, t, and covariates}

\item{inter}{a logical value indicating the existence of exposure-mediator interaction}

\item{p.U}{marginal probability of unmeasured confounder}

\item{outcome}{survival outcome}

\item{time}{survival time}

\item{m}{mediator}

\item{t}{exposure}

\item{covariates}{the vector of covariates}

\item{b.y, b.m, b.t}{the coefficients between U and exposure, mediator, and survival outcome}

\item{scale.m, scale.t}{scale of mediator and exposure, binary or continuous}

\item{iter_num}{number of iteration}
}
\value{
conditional probability of U
}
\description{
generate conditional probability of U
}
