% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{medsurvsens}
\alias{medsurvsens}
\title{estimate the NIE and NDE under each sensitivity parameter}
\usage{
medsurvsens(
  data,
  inter = FALSE,
  p.U = 0.5,
  outcome,
  time,
  m,
  t,
  covariates,
  scale.m,
  scale.t,
  range.b.y,
  range.b.m,
  range.b.t,
  ngrid = 10,
  iter_num = 20,
  multiple_draw = F,
  R = 5,
  B = 100,
  parallel = F
)
}
\arguments{
\item{data}{data frame including time, outcome, m, t, and covariates}

\item{inter}{a logical value indicating the existence of exposure-mediator interaction}

\item{p.U}{marginal probability of unmeasured confounder}

\item{outcome}{survival outcome}

\item{time}{survival time}

\item{m}{mediator}

\item{t}{exposure}

\item{covariates}{the vector of covariates}

\item{scale.m, scale.t}{scale of mediator and exposure, binary or continuous}

\item{range.b.y, range.b.m, range.b.t}{the range of coefficients between U and exposure, mediator, and survival outcome}

\item{ngrid}{defines how many intervals are used to divide range.b.y, range.b.m, and range.b.t into grid points.}

\item{iter_num}{number of iteration}

\item{multiple_draw}{whether to draw multiple samples of U}

\item{R}{the number of draws for multiple samples of U}

\item{B}{the number of bootstrap}

\item{parallel}{whether to use parallel computation}
}
\value{
data frame containing sensitivity parameters, the corresponding mediation effects, and p-values.
}
\description{
estimate the NIE and NDE under each sensitivity parameter
}
